// console/src/generated/QuotaWindowDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type QuotaWindowDto = { id: number | null, quota_id: number, subject_kind: string, subject_id: number, window_kind: string, window_start: number | null, reset_at: number | null, started: boolean, cost_used: string, cost_limit: string, };